Rename module context -> environment
