+++ packages/vm/src/environment.rs
