+++ packages/vm/src/lib.rs
@@ -3,8 +3,8 @@ mod cache;
 mod calls;
 mod checksum;
 mod compatibility;
-mod context;
 mod conversion;
+mod environment;
 mod errors;
 mod features;
 mod imports;
